+++ lib/assembled/widgets/cards/cards.dart
@@ -13,4 +13,5 @@ export 'baseComplete/baseComplete.dart';
 export 'editInterests/editInterests.dart';
 export 'referralCard/referralCard.dart';
 export 'shareOnSocial/shareOnSocial.dart';
+export 'planMeeting/planMeeting.dart';
 export 'planMeeting/filter/filter.dart';